test: make assertions on duplex.from instead of internal.duplexify
